# docs/api/openapi.yaml
openapi: 3.1.0
info: {title: FrameFlow API, version: 0.1.0}
servers: [{url: http://127.0.0.1:28741}]
security: [{bearerAuth: []}]
paths:
  /health:
    get: {security: [], responses: {'200': {description: Healthy}}}
  /api/projects:
    get: {responses: {'200': {description: Project list}}}
    post: {requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateProject'}}}}, responses: {'201': {description: Created}}}
  /api/projects/{id}:
    get: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Project}, '404': {description: Not found}}}
  /api/projects/{id}/drafts:
    post: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Draft saved}}}
  /api/tasks:
    get: {parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}], responses: {'200': {description: Task page}}}
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                kind: {type: string, enum: [story, image, video]}
                prompt: {type: string}
                model_id: {type: string, description: Model selected from a provider connection; defaults to the operation default when omitted}
                aspect_ratio: {type: string}
                source_image_url: {type: string}
      responses: {'202': {description: Task accepted}}
  /api/tasks/{id}:
    get: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Task}}}
    delete: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'204': {description: Deleted}}}
  /api/tasks/{id}/cancel:
    post: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Cancelled}}}
  /api/tasks/{id}/retry:
    post: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'202': {description: Retried}}}
  /api/tasks/{id}/result:
    get: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Completed task metadata as a JSON attachment}, '409': {description: Result not ready}}}
  /api/tasks/{id}/download:
    get:
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: Saved PNG or MP4 attachment, content: {image/png: {schema: {type: string, format: binary}}, video/mp4: {schema: {type: string, format: binary}}}}
        '409': {description: No saved media file or result not ready}
  /ws:
    get:
      parameters:
        - {name: token, in: query, required: false, schema: {type: string}, description: API authentication token}
      responses:
        '101': {description: WebSocket broadcasts live task events to all connections; clients send text ping every 15 seconds and refresh tasks over HTTP after reconnect}
  /api/providers:
    get: {responses: {'200': {description: Provider list}}}
    post: {responses: {'201': {description: Provider saved}}}
  /api/providers/{id}/test:
    post: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Connection healthy}, '422': {description: Connection failed}}}
  /api/connections:
    get: {responses: {'200': {description: Vendor connections without secrets}}}
    post: {requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ConnectionInput'}}}}, responses: {'201': {description: Vendor connection created}}}
  /api/connections/{id}:
    put: {parameters: [{$ref: '#/components/parameters/ID'}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ConnectionInput'}}}}, responses: {'200': {description: Vendor connection updated}}}
    delete: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'204': {description: Deleted after models are removed}, '409': {description: Connection has models}}}
  /api/connections/{id}/test:
    post: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Model discovery endpoint reachable}, '422': {description: Authentication or discovery failed}}}
  /api/connections/{id}/models:
    get: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Models for this connection}}}
    post: {parameters: [{$ref: '#/components/parameters/ID'}], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ManualModel'}}}}, responses: {'201': {description: Model added; unknown models are not executable}}}
  /api/connections/{id}/models/sync:
    post: {parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Discovered models upserted without overwriting manual entries}}}
  /api/connections/{id}/models/{model}:
    patch: {parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/ModelID'}], responses: {'200': {description: Model enabled or selected as default}}}
    delete: {parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/ModelID'}], responses: {'204': {description: Model removed}}}
  /api/models:
    get: {parameters: [{name: capability, in: query, schema: {type: string, enum: [story, image, video]}}, {name: connection_id, in: query, schema: {type: string}}], responses: {'200': {description: Model list}}}
  /api/materials:
    get: {responses: {'200': {description: Material list}}}
    post: {responses: {'201': {description: Material saved}}}
components:
  securitySchemes: {bearerAuth: {type: http, scheme: bearer}}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string}}
    ModelID: {name: model, in: path, required: true, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
  schemas:
    CreateProject: {type: object, required: [name], properties: {name: {type: string, minLength: 1}}}
    ConnectionInput: {type: object, required: [id, name, vendor, base_url], properties: {id: {type: string}, name: {type: string}, vendor: {type: string, enum: [bailian]}, base_url: {type: string, format: uri, description: "Complete HTTPS Bailian API endpoint (for example, /compatible-mode/v1/chat/completions); other supported operations use the same workspace host with their documented endpoints"}, api_key: {type: string, writeOnly: true}}}
    ManualModel: {type: object, required: [model_id], properties: {model_id: {type: string}, name: {type: string}, capabilities: {type: array, items: {type: string, enum: [story, image, video]}}}}